Refactor actions out of ControlConnection
Move implementation of instance related actions out of ControlConnection
and into Controller to allow plugins to easily make use of these methods.
